refactored multiplayer and singleplayer state handling into separate components
